refactor: extract showWebView to ModalDisplayManager for better coverage
- Create ModalDisplayManager class to handle UI/modal display logic
- Update PayPalMessageDataProvider to delegate to ModalDisplayManager
- Add ModalDisplayManager to Kover exclusions (UI code)
- Update tests to mock ModalDisplayManager for better unit testing
