+++ library/src/main/java/com/paypal/messages/data/PayPalMessageDataProvider.kt
@@ -1,18 +1,11 @@
 package com.paypal.messages.data
 
 import android.content.Context
-import android.content.Intent
 import android.os.Handler
 import android.os.Looper
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.FragmentActivity
-import com.paypal.messages.ModalFragment
-import com.paypal.messages.PayPalModalActivity
 import com.paypal.messages.analytics.AnalyticsEvent
 import com.paypal.messages.analytics.EventType
 import com.paypal.messages.config.message.PayPalMessageConfig
-import com.paypal.messages.config.modal.ModalConfig
-import com.paypal.messages.config.modal.ModalEvents
 import com.paypal.messages.io.Api
 import com.paypal.messages.io.ApiMessageData
 import com.paypal.messages.io.ApiResult
@@ -20,7 +13,6 @@ import com.paypal.messages.io.OnActionCompleted
 import com.paypal.messages.utils.LogCat
 import com.paypal.messages.utils.PayPalErrors
 import java.util.UUID
-import java.util.WeakHashMap
 import java.util.concurrent.ConcurrentHashMap
 
 /**
@@ -108,12 +100,11 @@ interface PayPalMessageClickHandler {
  * }
  * ```
  */
-class PayPalMessageDataProvider {
+class PayPalMessageDataProvider(
+	private val modalDisplayManager: ModalDisplayManager = ModalDisplayManager(),
+) {
 	private val TAG = "PayPalMessageDataProvider"
 	private val handler = Handler(Looper.getMainLooper())
-	
-	// Track modal instances by instanceId to prevent leaks and ensure cleanup
-	private val modalInstances = WeakHashMap<UUID, ModalFragment>()
 
 	/**
 	 * Fetches message data using the provided configuration
@@ -230,23 +221,18 @@ class PayPalMessageDataProvider {
 						),
 					)
 	
-					// Show modal - make sure this is called to display the bottom sheet
+					// Show modal - delegate to ModalDisplayManager
 					LogCat.debug(TAG, "[$clickTraceId] Showing modal for instanceId: $instanceId")
 					
 					// Before showing the modal, clear any existing PayPalModalActivity with the same instance ID
 					try {
-						// Tell PayPalModalActivity to reset modal state
-						try {
-							com.paypal.messages.PayPalModalActivity.resetAllModals()
-							LogCat.debug(TAG, "[$clickTraceId] Reset all modal state")
-						} catch (e: Exception) {
-							LogCat.debug(TAG, "[$clickTraceId] Error resetting modal state: ${e.message}")
-						}
+						com.paypal.messages.PayPalModalActivity.resetAllModals()
+						LogCat.debug(TAG, "[$clickTraceId] Reset all modal state")
 					} catch (e: Exception) {
-						LogCat.debug(TAG, "[$clickTraceId] Error cleaning up fragments: ${e.message}")
+						LogCat.debug(TAG, "[$clickTraceId] Error resetting modal state: ${e.message}")
 					}
 					
-					showWebView(context, response, config, instanceId, onApply, onClick, onError)
+					modalDisplayManager.showModal(context, response, config, instanceId, onApply, onClick, onError)
 					LogCat.debug(TAG, "[$clickTraceId] Modal display request complete")
 				} catch (e: Exception) {
 					// Log any errors that occur
@@ -264,195 +250,8 @@ class PayPalMessageDataProvider {
 			}
 
 			override fun onCleanup() {
-				// Clean up modal if it exists
-				val modal = modalInstances[instanceId]
-				if (modal != null) {
-					try {
-						// Only dismiss if the fragment is still added to a fragment manager
-						// This prevents crashes during activity destruction
-						if (modal.isAdded) {
-							modal.dismiss()
-						}
-					} catch (e: IllegalStateException) {
-						// Fragment is no longer associated with a fragment manager
-						// This can happen during activity destruction - safe to ignore
-						LogCat.debug(TAG, "Fragment no longer associated with fragment manager during cleanup: ${e.message}")
-					} catch (e: Exception) {
-						// Log any other unexpected errors but don't crash
-						LogCat.error(TAG, "Error dismissing modal during cleanup: ${e.message}")
-					}
-					modalInstances.remove(instanceId)
-				}
-			}
-		}
-	}
-	
-	/**
-	 * Shows the web view modal
-	 */
-	private fun showWebView(
-		context: Context,
-		response: ApiMessageData.Response,
-		config: PayPalMessageConfig,
-		instanceId: UUID,
-		onApply: () -> Unit,
-		onClick: () -> Unit,
-		onError: (PayPalErrors.Base) -> Unit,
-	) {
-		LogCat.debug(TAG, "Showing modal with context: ${context.javaClass.simpleName}")
-		// Try to find a usable AppCompatActivity from the context
-		val appCompatContext = com.paypal.messages.utils.ContextCompatWrapper.findAppCompatActivity(context)
-		LogCat.debug(TAG, "Found AppCompatActivity: $appCompatContext")
-		
-		when {
-			// For JetpackActivity or other ComponentActivity, use PayPalModalActivity
-			context.javaClass.simpleName == "JetpackActivity" ||
-				(context is androidx.activity.ComponentActivity && context !is androidx.appcompat.app.AppCompatActivity) -> {
-				LogCat.debug(TAG, "Using PayPalModalActivity approach for ${context.javaClass.simpleName}")
-				
-				try {
-					// Launch the PayPalModalActivity directly
-					val intent = Intent(context, PayPalModalActivity::class.java).apply {
-						putExtra("CLIENT_ID", config.data.clientID)
-						putExtra("AMOUNT", config.data.amount)
-						putExtra("BUYER_COUNTRY", config.data.buyerCountry)
-						putExtra("OFFER_TYPE", response.meta?.offerType?.toString())
-						putExtra("INSTANCE_ID", instanceId.toString())
-						// Make sure it appears correctly
-						addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-						addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
-						// Force a new modal instance if needed
-						putExtra("FORCE_NEW", true)
-					}
-					
-					// Start the activity
-					context.startActivity(intent)
-					
-					// Register the callbacks - remove redundant onClick call to avoid double modal issue
-					PayPalModalActivity.registerCallbacks(
-						instanceId = instanceId,
-						onApply = onApply,
-						onClick = onClick,
-						onError = onError,
-					)
-					
-					LogCat.debug(TAG, "Successfully launched modal activity")
-				} catch (e: Exception) {
-					LogCat.error(TAG, "Failed to show modal: ${e.message}")
-					onError.invoke(PayPalErrors.ModalFailedToLoad("Failed to show modal: ${e.message}", null))
-				}
-			}
-			// For AppCompatActivity contexts, use ModalFragment
-			appCompatContext != null -> {
-				// Check if we have an existing modal and if it's still added to the fragment manager
-				// If it was dismissed, create a new one instead of reusing the dismissed fragment
-				val existingModal = modalInstances[instanceId]
-				val modal = if (existingModal != null && existingModal.isAdded) {
-					// Reuse existing modal that's still shown
-					existingModal
-				} else {
-					// Create a new modal (either doesn't exist or was dismissed)
-					// Remove the old reference if it exists but was dismissed
-					if (existingModal != null) {
-						modalInstances.remove(instanceId)
-					}
-					
-					val newModal = ModalFragment.newInstance(config.data.clientID)
-
-					// Build modal config
-					val modalConfig = ModalConfig(
-						amount = config.data.amount,
-						buyerCountry = config.data.buyerCountry,
-						offer = response.meta?.offerType,
-						ignoreCache = false,
-						devTouchpoint = false,
-						stageTag = null,
-						events = ModalEvents(
-							onApply = onApply,
-							// onClick is already called before this method is invoked - don't call it again
-							onClick = { /* onClick is handled earlier */ },
-							onError = onError,
-						),
-						modalCloseButton = response.meta?.modalCloseButton!!,
-					)
-
-					newModal.init(modalConfig)
-					newModal.show(appCompatContext.supportFragmentManager, newModal.tag)
-
-					// Store the modal instance
-					modalInstances[instanceId] = newModal
-
-					newModal
-				}
-
-				// modal.show() above will display the modal on initial view, but if the user closes the modal
-				// it will become visually hidden and this method will re-display the modal without
-				// attempting to reattach it
-				// the delay prevents noticeable shift when the offer type is changed
-				handler.postDelayed({
-					modal.expand()
-				}, 250)
-			}
-			
-			// For other ComponentActivity contexts (redundant now with the case above, but keeping for safety)
-			context is androidx.activity.ComponentActivity &&
-				context !is androidx.fragment.app.FragmentActivity &&
-				context !is androidx.appcompat.app.AppCompatActivity &&
-				context.javaClass.simpleName != "JetpackActivity" -> {
-				// For ComponentActivity, we'll need to delegate to an Activity-level function
-				// that can show a Compose-based modal
-				val intent = Intent(context, PayPalModalActivity::class.java).apply {
-					putExtra("CLIENT_ID", config.data.clientID)
-					putExtra("AMOUNT", config.data.amount)
-					putExtra("BUYER_COUNTRY", config.data.buyerCountry)
-					putExtra("OFFER_TYPE", response.meta?.offerType)
-					// We can't easily serialize the ModalCloseButton as a parcelable
-					// so we'll use the default in PayPalModalActivity
-					putExtra("INSTANCE_ID", instanceId.toString())
-					// Set flags to ensure it appears as a transparent overlay
-					addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-					addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
-					addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-				}
-				context.startActivity(intent)
-				
-				// Register the callbacks for the activity to use
-				PayPalModalActivity.registerCallbacks(
-					instanceId = instanceId,
-					onApply = onApply,
-					onClick = onClick,
-					onError = onError,
-				)
-			}
-			
-			else -> {
-				// Fall back to using the PayPalModalActivity approach with any context
-				// This ensures we can always show a modal
-				try {
-					val intent = Intent(context, PayPalModalActivity::class.java).apply {
-						putExtra("CLIENT_ID", config.data.clientID)
-						putExtra("AMOUNT", config.data.amount)
-						putExtra("BUYER_COUNTRY", config.data.buyerCountry)
-						putExtra("OFFER_TYPE", response.meta?.offerType)
-						putExtra("INSTANCE_ID", instanceId.toString())
-						// Set flags to ensure it appears as a transparent overlay
-						addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-						addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
-						addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-					}
-					context.startActivity(intent)
-					
-					// Register the callbacks for the activity to use
-					PayPalModalActivity.registerCallbacks(
-						instanceId = instanceId,
-						onApply = onApply,
-						onClick = onClick,
-						onError = onError,
-					)
-				} catch (e: Exception) {
-					LogCat.error(TAG, "Failed to show modal with context: ${context.javaClass.simpleName}: ${e.message}")
-					onError.invoke(PayPalErrors.UnsupportedContextException("Cannot show modal: ${e.message}"))
-				}
+				// Delegate cleanup to ModalDisplayManager
+				modalDisplayManager.cleanupModal(instanceId)
 			}
 		}
 	}
